requirements.txt
Django==6.0.3
djangorestframework==3.14.0
mssql-django==1.7
django-cors-headers==4.3.1
